Kiosk/requirements.txt
absl-py==2.0.0
altair==5.1.2
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.7.0
build==1.0.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
comtypes==1.3.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.5.0
decorator==5.1.0
deepface==0.0.81
Deprecated==1.2.14
entrypoints==0.3
exceptiongroup==1.2.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
Flask-Cors==4.0.0
flask-ngrok==0.0.25
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
fonttools==4.46.0
gast==0.5.4
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
imutils==0.5.4
iniconfig==2.0.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-client==7.0.6
jupyter-core==4.8.1
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.3
mdurl==0.1.2
ml-dtypes==0.2.0
mtcnn==0.1.1
mysql-connector-python==8.3.0
nest-asyncio==1.5.1
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.2
parso==0.8.2
pickleshare==0.7.5
Pillow==10.1.0
pluggy==1.3.0
prompt-toolkit==3.0.20
protobuf==4.23.4
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
pypiwin32==223
pyproject_hooks==1.0.0
PySocks==1.7.1
pytest==7.4.3
python-dateutil==2.8.2
python-engineio==4.9.0
python-socketio==5.11.1
pyttsx3==2.90
pytz==2023.3.post1
pywin32==302
pyzmq==22.3.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rich==13.6.0
rpds-py==0.12.0
rsa==4.9
scipy==1.12.0
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.28.1
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.1
tqdm==4.66.1
traitlets==5.1.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.5
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
zipp==3.17.0
